% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duck.R
\name{duck}
\alias{duck}
\title{Replace all errors}
\usage{
duck(polite = TRUE)
}
\arguments{
\item{polite}{`logical(1)`. If `TRUE` returns `Error: duck` on error, otherwise returns something a little more...
sweary.}
}
\description{
Replace all errors
}
\examples{
\dontrun{
duck()
duck(polite = FALSE)
}

}
